Renamed proto_version -> version_prefix
